backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
fastapi-cors==0.0.6
pydantic==2.5.0
python-dateutil==2.8.2
numpy==1.24.3